+++ NEWS.md
@@ -1,3 +1,9 @@
+# klassR 0.2.1
+
+- Added option to fetch correspondence table using table ID only.
+- Added quiet option to print/or suppress the url address which the class is fetched from.
+
+
 # klassR 0.2.0
 
 - `GetKlass()` now includes option to fetch notes